+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes that have landed in master but are not yet released
 
 * Add license field to all packages 
 
+### Form
+
+* Fix - when term supplied without indicator (such as excludeTerm and excluded), indicator first true until calculated
+
+* Fix - remove evaluate state changes from init field - its already evaluated in 'evaluateField'
+
 ### React Layout
 
 * Footer popover - placement in parent container instead of body